Remove menu folder
